backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
psycopg2-binary
PyJWT
groq
pandas
openpyxl
supabase
